remove tox3 support
