+++ azure-pipelines.yml
@@ -12,8 +12,6 @@ jobs:
         python.version: '3.9'
       Python38:
         python.version: '3.8'
-      Python37:
-        python.version: '3.7'
     maxParallel: 5
 
   steps:
@@ -30,12 +28,3 @@ jobs:
 
   - script: ./ci.sh tox-4.x docker-4.x
     displayName: 'Tox 4.x, Docker 4.x'
-
-  - script: ./ci.sh tox-3.x docker-6.x
-    displayName: 'Tox 3.x, Docker 6.x'
-
-  - script: ./ci.sh tox-3.x docker-5.x
-    displayName: 'Tox 3.x, Docker 5.x'
-
-  - script: ./ci.sh tox-3.x docker-4.x
-    displayName: 'Tox 3.x, Docker 4.x'
