+++ ci.sh
@@ -5,12 +5,6 @@ set -x
 tox_version=$1
 docker_version=$2
 
-if [ "$tox_version" = "tox-4.x" ]; then
-    mypy_target=mypy-tox4
-else
-    mypy_target=mypy-tox3
-fi
-
 
 # our tox.ini sets up volumes within .tox, so make sure it exists
 # before trying to run tox...
@@ -24,7 +18,7 @@ python tox_docker/tests/capture_containers_and_volumes.py .
 tox -e integration
 python tox_docker/tests/assert_containers_and_volumes_unchanged.py .
 
-tox -e $mypy_target
+tox -e mypy
 
 echo "testing health check failure handling, an ERROR is expected:"
 tox -e healthcheck-failing 2>&1 | grep "'toxdocker/healthcheck' (from 'healthcheck-failing') failed health check"
