+++ setup.cfg
@@ -13,7 +13,7 @@ force_grid_wrap = false
 from_first = true
 include_trailing_comma = true
 # known_third_party is autogenereated. Do not edit.
-known_third_party =docker,packaging,py,pytest,setuptools,tox,vcversioner
+known_third_party =docker,pytest,setuptools,tox,vcversioner
 multi_line_output = 3
 not_skip = __init__.py
 order_by_type = false
