+++ setup.py
@@ -2,15 +2,14 @@ from setuptools import find_packages, setup
 
 setup(
     name="tox-docker",
-    description="Launch a docker instance around test runs",
+    description="Manage lifecycle of docker containers during Tox test runs",
     long_description=open("README.rst").read(),
     url="https://github.com/tox-dev/tox-docker",
     maintainer="Dan Crosta",
     maintainer_email="dcrosta@late.am",
     install_requires=[
         "docker>=4.0,<7.0",
-        "packaging",
-        "tox>=3.0.0,<5.0",
+        "tox>=4.0.0,<5.0",
     ],
     packages=find_packages(),
     entry_points={"tox": ["docker = tox_docker"]},
@@ -20,10 +19,10 @@ setup(
         "Environment :: Plugins",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: BSD License",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Topic :: Software Development :: Testing",
     ],
 )
