--- tox-3.x
@@ -1,1 +1,0 @@
-tox>=3.0,<4.0
