--- tox-4.x
@@ -1,1 +1,0 @@
---pre tox>=4.0.0b2,<5.0
