+++ tox.ini
@@ -1,5 +1,7 @@
 [tox]
-envlist = integration
+envlist =
+    integration
+    mypy
 
 [testenv:docs]
 deps =
@@ -29,19 +31,12 @@ commands =
     py.test [] {toxinidir}/tox_docker
     python -c 'import os; os.remove(os.environ["VOLUME_DIR"] + "/healthy")'
 
-[testenv:mypy-tox3]
+[testenv:mypy]
 skip_install = true
 deps =
     mypy==0.910
 commands =
-    mypy --exclude tox4 {toxinidir}/tox_docker
-
-[testenv:mypy-tox4]
-skip_install = true
-deps =
-    mypy==0.910
-commands =
-    mypy --exclude tox3 {toxinidir}/tox_docker
+    mypy {toxinidir}/tox_docker
 
 [docker:networking-one]
 image = ksdn117/tcp-udp-test
