--- tox_docker/log.py
@@ -1,3 +1,0 @@
-from typing import Callable
-
-LogFunc = Callable[[str], None]
