+++ tox_docker/plugin.py
@@ -1,3 +1,11 @@
+__all__ = (
+    "tox_add_env_config",
+    "tox_add_option",
+    "tox_after_run_commands",
+    "tox_before_run_commands",
+)
+
+from logging import getLogger
 from typing import Dict, Iterable, List, Mapping, Optional, Tuple, Union
 import os
 import socket
@@ -6,10 +14,26 @@ import time
 
 from docker.errors import ImageNotFound, NotFound
 from docker.models.containers import Container
+from tox.config.cli.parser import ToxParser
+from tox.config.loader.section import Section
+from tox.config.sets import EnvConfigSet
+from tox.execute.api import Outcome
+from tox.plugin import impl
+from tox.session.state import State
+from tox.tox_env.api import ToxEnv
+from tox.tox_env.errors import Fail
 import docker as docker_module
 
-from tox_docker.config import ContainerConfig, RunningContainers
-from tox_docker.log import LogFunc
+from tox_docker.config import (
+    ContainerConfig,
+    DockerConfigSet,
+    parse_container_config,
+    RunningContainers,
+)
+
+
+def log(line: str) -> None:
+    getLogger().warning(f"docker> {line}")
 
 
 class HealthCheckFailed(Exception):
@@ -52,14 +76,14 @@ def escape_env_var(varname: str) -> str:
     return "".join(varletters)
 
 
-def docker_build_or_pull(container_config: ContainerConfig, log: LogFunc) -> None:
+def docker_build_or_pull(container_config: ContainerConfig) -> None:
     if container_config.image:
-        docker_pull(container_config, log)
+        docker_pull(container_config)
     else:
-        docker_build(container_config, log)
+        docker_build(container_config)
 
 
-def docker_pull(container_config: ContainerConfig, log: LogFunc) -> None:
+def docker_pull(container_config: ContainerConfig) -> None:
     assert container_config.image
 
     docker = docker_module.from_env(version="auto")
@@ -73,7 +97,7 @@ def docker_pull(container_config: ContainerConfig, log: LogFunc) -> None:
     container_config.runnable_image = docker.images.get(str(container_config.image))
 
 
-def docker_build(container_config: ContainerConfig, log: LogFunc) -> None:
+def docker_build(container_config: ContainerConfig) -> None:
     assert container_config.dockerfile
 
     docker = docker_module.from_env(version="auto")
@@ -100,7 +124,6 @@ def docker_build(container_config: ContainerConfig, log: LogFunc) -> None:
 def docker_run(
     container_config: ContainerConfig,
     running_containers: RunningContainers,
-    log: LogFunc,
 ) -> Container:
     docker = docker_module.from_env(version="auto")
 
@@ -152,7 +175,7 @@ def docker_run(
 
 
 def docker_health_check(
-    container_config: ContainerConfig, container: Container, log: LogFunc
+    container_config: ContainerConfig, container: Container
 ) -> None:
     docker = docker_module.from_env(version="auto")
 
@@ -171,9 +194,7 @@ def docker_health_check(
                 raise HealthCheckFailed(msg)
 
 
-def docker_stop(
-    container_config: ContainerConfig, container: Container, log: LogFunc
-) -> None:
+def docker_stop(container_config: ContainerConfig, container: Container) -> None:
     if container_config.stop:
         log(f"remove '{container.short_id}' (from {container_config.name!r})")
         container.remove(v=True, force=True)
@@ -189,11 +210,9 @@ def docker_get(container_config: ContainerConfig) -> Optional[Container]:
         return None
 
 
-def stop_containers(
-    containers: Iterable[Tuple[ContainerConfig, Container]], log: LogFunc
-) -> None:
+def stop_containers(containers: Iterable[Tuple[ContainerConfig, Container]]) -> None:
     for container_config, container in containers:
-        docker_stop(container_config, container, log)
+        docker_stop(container_config, container)
 
 
 def get_env_vars(
@@ -217,3 +236,103 @@ def get_env_vars(
         env[escape_env_var(f"{container_config.name}_{containerport}_PORT")] = hostport
 
     return env
+
+
+@impl
+def tox_add_env_config(env_conf: EnvConfigSet, state: State) -> None:
+    def build_docker_config_set(container_name: object) -> DockerConfigSet:
+        assert isinstance(container_name, str)
+        docker_conf = state.conf.get_section_config(
+            section=Section("docker", container_name),
+            base=[],
+            of_type=DockerConfigSet,
+            for_env=None,
+        )
+        if not docker_conf.loaders:
+            raise ValueError(f"Missing [docker:{container_name}] in tox.ini")
+        return docker_conf
+
+    env_conf.add_config(
+        keys=["docker"],
+        of_type=List[DockerConfigSet],
+        default=[],
+        desc="docker image configs to load",
+        factory=build_docker_config_set,  # type: ignore
+    )
+
+
+@impl
+def tox_before_run_commands(tox_env: ToxEnv) -> None:
+    docker_confs = tox_env.conf.load("docker")
+
+    container_configs = [
+        parse_container_config(docker_conf) for docker_conf in docker_confs
+    ]
+
+    seen = set()
+    for container_config in container_configs:
+        if container_config.name in seen:
+            raise ValueError(
+                f"Container {container_config.name!r} specified more than once"
+            )
+        seen.add(container_config.name)
+
+    for container_config in container_configs:
+        docker_build_or_pull(container_config)
+
+    config_and_container = []
+    running_containers: RunningContainers = {}
+    for container_config in container_configs:
+        container = docker_run(container_config, running_containers)
+        config_and_container.append((container_config, container))
+        running_containers[container_config.runas_name] = container
+
+    for container_config, container in config_and_container:
+        try:
+            docker_health_check(container_config, container)
+        except HealthCheckFailed:
+            tox_env.interrupt()
+            clean_up_containers(tox_env)
+            raise Fail(
+                f"{container_config.image!r} (from {container_config.name!r}) failed health check"
+            )
+
+        tox_env.conf["set_env"].update(get_env_vars(container_config, container))
+
+
+@impl
+def tox_after_run_commands(
+    tox_env: ToxEnv, exit_code: int, outcomes: List[Outcome]
+) -> None:
+    clean_up_containers(tox_env)
+
+
+def clean_up_containers(tox_env: ToxEnv) -> None:
+    docker_confs = tox_env.conf.load("docker")
+
+    container_configs = [
+        parse_container_config(docker_conf) for docker_conf in docker_confs
+    ]
+
+    configs_and_containers = []
+    for config in container_configs:
+        container = docker_get(config)
+        if container:
+            configs_and_containers.append((config, container))
+
+    stop_containers(configs_and_containers)
+
+
+@impl
+def tox_add_option(parser: ToxParser) -> None:
+    # command line flag to keep docker containers running
+    parser.add_argument(
+        "--docker-dont-stop",
+        default=[],
+        action="append",
+        metavar="CONTAINER",
+        help=(
+            "If specified, tox-docker will not stop CONTAINER after the test run. "
+            "Can be specified multiple times."
+        ),
+    )
