--- tox_docker/tox4/log.py
@@ -1,5 +1,0 @@
-from logging import getLogger
-
-
-def log(line: str) -> None:
-    getLogger().warning(f"docker> {line}")
